Modify ConnectSensoAPI so that it takes in user input values
DETAILED DESCRIPTION
--------------------
I have modified ConnectSensoAPI.java so that it takes in a Hashmap of values that can then be parsed into a JSON object to be sent to the senso endpoint. I have also renamed some files so that it describes what the class does better.

TESTING INFORMATION
-------------------
There is a new test class called ConnectSensoAPITest.java, which test whether we can create a mapping and use that mapping as our request body to send with our senso api call.