Remove and refactor duplicate formatting files
DETAILED DESCRIPTION
--------------------
Remove the SensoEndpointHandler because we don't need that endpoint anymore (its role is covered by the ServerMainEndpointHandler). I have also removed the duplicate formatting files.

TESTING INFORMATION
-------------------
This change is not tested, because it is just removing files that are no longer needed (no other classes use them).