Improve the formatting for POST response
DETAILED DESCRIPTION
--------------------
Modified ServerMainEndpointHandler to return a response body that is nicer to parse for the front-end. Complete documentation for the class, as well as other related classes like RootEndpointHandler and RestServiceCorsApplications.

TESTING INFORMATION
-------------------
Created and completed a full test suit for ServerMainEndpointHandler, as well as manually testing the result using postman.