Handle exceptions for the server endpoint
DETAILED DESCRIPTION
--------------------
Created a DatabaseConnectionFailureException and a SensoConnectionFailureException classes to handle all exceptions when connecting to the outer entities. I have made it so that if these two exceptions are thrown, the server will return a string containing error message instead of the json containing relevant information.

TESTING INFORMATION
-------------------
Created the exception for the exception class, and test that exception can be called and returns the correct error messages.